requirements.txt
py-clob-client>=0.34.6
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.0.0
eth-account>=0.10.0
rich>=14.0.0
flask>=3.0.3,<3.1.0
flask-socketio>=5.6.0
websockets>=12.0.0
